implemented collection view to begin from center and hold tap note to show options

+++ Keey/DrumViewModel.h
@@ -26,7 +26,7 @@ typedef enum DrumMood : NSUInteger {
 } DrumMood;
 
 - (void) setupDrumIntruments : (int)steps;
-- (void) updateStepAt : (int)index forInstrument :(enum DrumMood) drumMood;
-- (BOOL) shouldPlaySoundAt : (int)index forInstrument :(enum DrumMood)drumMood;
+- (void) updateStepAt : (NSUInteger)index forInstrument :(enum DrumMood) drumMood;
+- (BOOL) shouldPlaySoundAt : (NSUInteger)index forInstrument :(enum DrumMood)drumMood;
 
 @end
